# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kikae",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/tailor-api/public/api",
            "description": "Kikae Server #1"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/getTokens": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "getTokens"
                ]
            }
        },
        "/getUser": {
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getUser"
                ]
            }
        },
        "/currentUser": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "currentUser"
                ]
            },
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "currentUser"
                ]
            }
        },
...
        "/getUser/{id}": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "getUser"
                ]
            }
        },
        "/adminUpdateUser": {
            "post": {
                "summary": "Update the specified resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "adminUpdateUser"
                ]
            }
        },
        "/getStores": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getStores"
                ]
            },
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getStores"
                ]
            }
        },
        "/getStores/{user_id}": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "getStores"
                ]
            }
        },
        "/getProducts": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getProducts"
                ]
            },
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getProducts"
                ]
            }
        },
        "/getProduct/{id}": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "getProduct"
                ]
            }
        },
        "/getStoreProducts/{store_id}": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "getStoreProducts"
                ]
            }
        },
        "/getStoreProducts": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getStoreProducts"
                ]
            }
        },
        "/getStoreLikes": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getStoreLikes"
                ]
            }
        },
        "/getProductLikes": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getProductLikes"
                ]
            }
        },
        "/likeStore": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "likeStore"
                ]
            }
        },
        "/likeProduct": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "likeProduct"
                ]
            }
        },
        "/getStoreRating": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getStoreRating"
                ]
            }
        },
        "/rateStore": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "rateStore"
                ]
            }
        },
        "/getProductRating": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getProductRating"
                ]
            }
        },
        "/rateProduct": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "rateProduct"
                ]
            }
        },
        "/adminUpdateProduct": {
            "post": {
                "summary": "Update the specified resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "adminUpdateProduct"
                ]
            }
        },
        "/addProduct": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "addProduct"
                ]
            }
        },
        "/deleteProduct": {
            "post": {
                "summary": "Remove the specified resource from storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "deleteProduct"
                ]
            }
        },
        "/addProductMedia": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "addProductMedia"
                ]
            }
        },
        "/createBoomerang": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "createBoomerang"
                ]
            }
        },
        "/getBoomerangs": {
            "post": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "getBoomerangs"
                ]
            }
        },
...
    },
...
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}